# Refactor and optimize 'In'. Add benchmark.