@nuss-justin 2015-08-08
Added 2 more optimizations and updated benchmarks.