@jmoiron 2015-08-11
Thanks so much!  Great change :)